fixed PR challenges
